Seperate daily reports into its own component and reuse for patient and physician side
