+++ react_frontend/src/components/DailyReportComponent.css
@@ -1,0 +1,142 @@
+.daily-reports-content {
+  margin-top: 100px;
+  margin-left: 80px;
+  margin-right: 80px;
+  text-align: left;
+  padding-bottom: 50px;
+  width: -webkit-fill-available;
+}
+
+.daily-reports-content h1 {
+  font-size: 50px;
+  font-weight: bold;
+}
+
+.patient-info-box {
+  border: 4px solid #516a80;
+  background-color: #e7eaf8;
+  color: #2f3b66;
+  padding: 20px;
+  margin-top: -10px;
+  padding-bottom: 10px;
+  border-radius: 22px;
+}
+
+.patient-info-box h2 {
+  margin-top: 0;
+}
+
+.header-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.date-filter {
+  text-align: center;
+}
+
+.date-filter select {
+  border-radius: 20px;
+  font-size: 18px;
+  padding: 8px 12px;
+  border: 2px solid #516a80;
+  background-color: #e7eaf8;
+  color: #333;
+}
+
+.test-results-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.test-result-box {
+  border: 3px solid #516a80;
+  background-color: #e7eaf8;
+  padding: 15px;
+  border-radius: 15px;
+  width: calc(50% - 10px);
+  text-align: left;
+}
+
+.test-result-box h3 {
+  margin-top: 0;
+  font-size: 22px;
+}
+
+.game-title {
+  text-align: left;
+  font-size: 30px;
+  margin-bottom: 8px;
+  font-weight: bold;
+}
+
+.game-box {
+  padding-left: 0px;
+  margin-bottom: 60px;
+}
+
+.game-metrics {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 35px;
+  justify-content: center;
+}
+
+.metric-box {
+  background-color: #e7eaf8;
+  border: 2px solid #2f3b66;
+  margin-top: 20px;
+  margin-bottom: 0px;
+  padding: 10px;
+  border-radius: 40px;
+  text-align: center;
+  width: 175px;
+  box-shadow: 10px 10px 1px #2f3b66;
+  white-space: normal;
+  word-wrap: break-word;
+}
+
+.metric-box h4 {
+  font-size: 16px;
+  margin-bottom: 5px;
+  font-weight: bold;
+}
+
+.metric-box p {
+  font-size: 35px;
+  font-weight: bold;
+  margin-top: 7px;
+  margin-bottom: 7px;
+}
+
+.comparison {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.comparison-text {
+  font-size: 0.9em;
+  color: #555;
+  margin-top: 5px;
+  margin-bottom: 5px;
+}
+
+.comparison-top {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.arrow-icon {
+  width: 16px;
+  height: 16px;
+}
+
+.percentage-change {
+  font-weight: bold;
+}
