+++ react_frontend/src/pages/patient/PatientReportPage.js
@@ -1,0 +1,46 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import titleImage from "../../assets/title.svg";
+import homeIcon from "../../assets/home-light.svg";
+import reportIcon from "../../assets/reports-dark.svg";
+import supportIcon from "../../assets/support-light.svg";
+import profileIcon from "../../assets/profile-light.svg";
+import DailyReportComponent from "../../components/DailyReportComponent";
+import "./PatientReportPage.css";
+
+const PatientReportPage = () => {
+  return (
+    <div className="patient-home-container">
+      {/* Left Navigation */}
+      <div className="left-side">
+        <div className="title-container">
+          <img src={titleImage} alt="Title" className="title-image" />
+        </div>
+        <div className="menu">
+          <Link to="/patient-home-page" className="menu-item link">
+            <img src={homeIcon} alt="Home" />
+            <span className="home-text">Home</span>
+          </Link>
+          <Link to="/patient-report-page" className="menu-item link">
+            <img src={reportIcon} alt="My Reports" />
+            <span style={{ color: "#2F3B66" }}>My Reports</span>
+          </Link>
+          <Link to="/support-page" className="menu-item link">
+            <img src={supportIcon} alt="Support" />
+            <span style={{ color: "#516A80" }}>Support</span>
+          </Link>
+        </div>
+        <Link to="/patient-login" className="menu-item link logout">
+          <img src={profileIcon} alt="Log Out" />
+          <span>Log Out</span>
+        </Link>
+      </div>
+      {/* Right Side: Shared Daily Report Content */}
+      <div className="right-side">
+        <DailyReportComponent />
+      </div>
+    </div>
+  );
+};
+
+export default PatientReportPage;
