--- react_frontend/src/pages/patient/PatientReportingLandingPage.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import './PatientReportingLandingPage.css';
-
-const PatientReportLandingPage = () => {
-  return (
-    <div className="patient-reporting-landing-container">
-      <h1>Patient Report Landing Page</h1>
-      <p>This is the Patient Report Landing Page.</p>
-    </div>
-  );
-};
-
-export default PatientReportLandingPage;
